평점 출력 오류 수정 및 불필요한 데이터 제거

+++ src/RatingManager.cpp
@@ -55,7 +55,7 @@ void RatingManager::loadFromFile(const std::string& filename) {
         int movieId = std::stoi(token);
 
         std::getline(ss, token, ',');
-        int score = std::stoi(token);
+        double score = std::stod(token);
 
         ratings.push_back(Rating(userId, movieId, score));
     }
